Move credentials to their own file (#627)
