+++ dbt/adapters/databricks/__init__.py
@@ -1,7 +1,7 @@
 from dbt.adapters.base import AdapterPlugin
 from dbt.adapters.databricks.column import DatabricksColumn  # noqa
 from dbt.adapters.databricks.connections import DatabricksConnectionManager  # noqa
-from dbt.adapters.databricks.connections import DatabricksCredentials
+from dbt.adapters.databricks.credentials import DatabricksCredentials
 from dbt.adapters.databricks.impl import DatabricksAdapter
 from dbt.adapters.databricks.relation import DatabricksRelation  # noqa
 from dbt.include import databricks
