+++ tests/unit/python/test_python_submissions.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import Mock
 
-from dbt.adapters.databricks.connections import DatabricksCredentials
+from dbt.adapters.databricks.credentials import DatabricksCredentials
 from dbt.adapters.databricks.python_submissions import BaseDatabricksHelper
 from dbt.adapters.databricks.python_submissions import DBContext
