+++ tests/unit/test_adapter.py
@@ -5,18 +5,17 @@ from typing import Dict
 from typing import Optional
 from unittest import mock
 
-import pytest
-from agate import Row
-
 import dbt.adapters.exceptions
 import dbt.flags as flags
+import pytest
+from agate import Row
 from dbt.adapters.databricks import __version__
 from dbt.adapters.databricks import DatabricksAdapter
 from dbt.adapters.databricks import DatabricksRelation
 from dbt.adapters.databricks.column import DatabricksColumn
-from dbt.adapters.databricks.connections import CATALOG_KEY_IN_SESSION_PROPERTIES
-from dbt.adapters.databricks.connections import DBT_DATABRICKS_HTTP_SESSION_HEADERS
-from dbt.adapters.databricks.connections import DBT_DATABRICKS_INVOCATION_ENV
+from dbt.adapters.databricks.credentials import CATALOG_KEY_IN_SESSION_PROPERTIES
+from dbt.adapters.databricks.credentials import DBT_DATABRICKS_HTTP_SESSION_HEADERS
+from dbt.adapters.databricks.credentials import DBT_DATABRICKS_INVOCATION_ENV
 from dbt.adapters.databricks.impl import check_not_found_error
 from dbt.adapters.databricks.impl import get_identifier_list_string
 from dbt.config import RuntimeConfig
