+++ tests/unit/test_compute_config.py
@@ -2,6 +2,7 @@ import unittest
 
 import dbt.exceptions
 from dbt.adapters.databricks import connections
+from dbt.adapters.databricks.credentials import DatabricksCredentials
 from dbt.contracts.graph import model_config
 from dbt.contracts.graph import nodes
 
@@ -15,7 +16,7 @@ class TestDatabricksConnectionHTTPPath(unittest.TestCase):
 
     def test_get_http_path_model(self):
         default_path = "my_http_path"
-        creds = connections.DatabricksCredentials(http_path=default_path)
+        creds = DatabricksCredentials(http_path=default_path)
 
         path = connections._get_http_path(None, creds)
         self.assertEqual(default_path, path)
@@ -72,7 +73,7 @@ class TestDatabricksConnectionHTTPPath(unittest.TestCase):
 
     def test_get_http_path_seed(self):
         default_path = "my_http_path"
-        creds = connections.DatabricksCredentials(http_path=default_path)
+        creds = DatabricksCredentials(http_path=default_path)
 
         path = connections._get_http_path(None, creds)
         self.assertEqual(default_path, path)
@@ -129,7 +130,7 @@ class TestDatabricksConnectionHTTPPath(unittest.TestCase):
 
     def test_get_http_path_snapshot(self):
         default_path = "my_http_path"
-        creds = connections.DatabricksCredentials(http_path=default_path)
+        creds = DatabricksCredentials(http_path=default_path)
 
         path = connections._get_http_path(None, creds)
         self.assertEqual(default_path, path)
